bridge/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
pyserial==3.5
# Use headless build for server environments (no GUI). On Raspberry Pi prefer system OpenCV where possible.
opencv-python-headless>=4.8.0.76
onnxruntime>=1.20.0
numpy>=2.1.2
